game-flow: move common bits to libtrx
