+++ data/tr1/ship/cfg/TR1X_gameflow_ub.json5
@@ -36,7 +36,7 @@
             "data/injections/pda_model.bin",
         ],
         "sequence": [
-            {"type": "display_picture", "path": "data/images/eidos.webp", "display_time": 1},
+            {"type": "display_picture", "path": "data/images/eidos.webp", "display_time": 1, "fade_in_time": 1.0, "fade_out_time": 1.0},
             {"type": "play_fmv", "fmv_id": 0},
             {"type": "play_fmv", "fmv_id": 1},
             {"type": "exit_to_title"},
@@ -110,11 +110,11 @@
                 {"type": "play_level"},
                 {"type": "level_stats"},
                 {"type": "play_music", "music_track": 19},
-                {"type": "display_picture", "path": "data/images/end.webp", "display_time": 7.5},
-                {"type": "display_picture", "path": "data/images/credits_ub.webp", "display_time": 7.5},
-                {"type": "display_picture", "path": "data/images/credits_1.webp", "display_time": 7.5},
-                {"type": "display_picture", "path": "data/images/credits_2.webp", "display_time": 7.5},
-                {"type": "display_picture", "path": "data/images/credits_3.webp", "display_time": 7.5},
+                {"type": "display_picture", "path": "data/images/end.webp", "display_time": 7.5, "fade_in_time": 1.0, "fade_out_time": 1.0},
+                {"type": "display_picture", "path": "data/images/credits_ub.webp", "display_time": 7.5, "fade_in_time": 1.0, "fade_out_time": 1.0},
+                {"type": "display_picture", "path": "data/images/credits_1.webp", "display_time": 7.5, "fade_in_time": 1.0, "fade_out_time": 1.0},
+                {"type": "display_picture", "path": "data/images/credits_2.webp", "display_time": 7.5, "fade_in_time": 1.0, "fade_out_time": 1.0},
+                {"type": "display_picture", "path": "data/images/credits_3.webp", "display_time": 7.5, "fade_in_time": 1.0, "fade_out_time": 1.0},
                 {"type": "total_stats", "background_path": "data/images/install.webp"},
                 {"type": "exit_to_title"},
             ],
