+++ src/libtrx/meson.build
@@ -113,6 +113,7 @@ sources = [
   'game/game_flow/common.c',
   'game/game_flow/reader.c',
   'game/game_flow/sequencer.c',
+  'game/game_flow/sequencer_misc.c',
   'game/game_flow/vars.c',
   'game/game_string.c',
   'game/game_string_table/common.c',
