+++ src/tr1/game/level.h
@@ -2,5 +2,5 @@
 
 #include "game/game_flow/types.h"
 
-void Level_Load(const GF_LEVEL *level);
 bool Level_Initialise(const GF_LEVEL *level);
+void Level_Load(const GF_LEVEL *level);
