+++ src/tr1/global/enum_map.def
@@ -33,45 +33,3 @@ ENUM_MAP_DEFINE(TARGET_LOCK_MODE, TLM_NONE, "no-lock")
 ENUM_MAP_DEFINE(MUSIC_LOAD_CONDITION, MUSIC_LOAD_NEVER, "never")
 ENUM_MAP_DEFINE(MUSIC_LOAD_CONDITION, MUSIC_LOAD_NON_AMBIENT, "non-ambient")
 ENUM_MAP_DEFINE(MUSIC_LOAD_CONDITION, MUSIC_LOAD_ALWAYS, "always")
-
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LOAD_LEVEL,        "load_level")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_LEVEL,        "play_level")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_FMV,          "play_fmv")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_CUTSCENE,     "play_cutscene")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_MUSIC,        "play_music")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LOADING_SCREEN,    "loading_screen")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_DISPLAY_PICTURE,   "display_picture")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LEVEL_STATS,       "level_stats")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_TOTAL_STATS,       "total_stats")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_EXIT_TO_TITLE,     "exit_to_title")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LEVEL_COMPLETE,    "level_complete")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_SET_CAMERA_ANGLE,  "set_cutscene_angle")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_FLIP_MAP,          "flip_map")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_WEAPONS,    "remove_weapons")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_SCIONS,     "remove_scions")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_AMMO,       "remove_ammo")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_MEDIPACKS,  "remove_medipacks")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_ADD_ITEM,          "give_item")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_MESH_SWAP,         "mesh_swap")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_SETUP_BACON_LARA,  "setup_bacon_lara")
-
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_TITLE, "title")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_NORMAL, "normal")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_CUTSCENE, "cutscene")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_GYM, "gym")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_BONUS, "bonus")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_DUMMY, "dummy")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_CURRENT, "current")
-
-ENUM_MAP_DEFINE(GF_ACTION, GF_NOOP, "noop")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_GAME, "play_level")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_SAVED_GAME, "load_saved_game")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_CINE, "play_cutscene")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_DEMO, "play_demo")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_FMV, "play_fmv")
-ENUM_MAP_DEFINE(GF_ACTION, GF_EXIT_TO_TITLE, "exit_to_title")
-ENUM_MAP_DEFINE(GF_ACTION, GF_LEVEL_COMPLETE, "level_complete")
-ENUM_MAP_DEFINE(GF_ACTION, GF_EXIT_GAME, "exit_game")
-ENUM_MAP_DEFINE(GF_ACTION, GF_RESTART_GAME, "restart_level")
-ENUM_MAP_DEFINE(GF_ACTION, GF_SELECT_GAME, "select_level")
-ENUM_MAP_DEFINE(GF_ACTION, GF_STORY_SO_FAR, "story_so_far")
