+++ src/tr2/game/level.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "game/game_flow/types.h"
-#include "global/types.h"
+
+#include <libtrx/game/level/common.h>
 
 bool Level_Initialise(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
 bool Level_Load(const GF_LEVEL *level);
