+++ src/tr2/global/enum_map.def
@@ -12,38 +12,3 @@ ENUM_MAP_DEFINE(RENDER_MODE, RM_HARDWARE, "hardware")
 ENUM_MAP_DEFINE(ASPECT_MODE, AM_ANY, "any")
 ENUM_MAP_DEFINE(ASPECT_MODE, AM_4_3, "4:3")
 ENUM_MAP_DEFINE(ASPECT_MODE, AM_16_9, "16:9")
-
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_EXIT_TO_TITLE,     "exit_to_title")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_DISPLAY_PICTURE,   "display_picture")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_ENABLE_SUNSET,     "enable_sunset")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_WEAPONS,    "remove_weapons")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_REMOVE_AMMO,       "remove_ammo")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LEVEL_COMPLETE,    "level_complete")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_LEVEL_STATS,       "level_stats")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_TOTAL_STATS,       "total_stats")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_SET_NUM_SECRETS,   "set_secret_count")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_SET_CAMERA_ANGLE,  "set_cutscene_angle")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_SET_START_ANIM,    "set_lara_start_anim")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_LEVEL,        "play_level")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_CUTSCENE,     "play_cutscene")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_FMV,          "play_fmv")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_DISABLE_FLOOR,     "disable_floor")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_ADD_ITEM,          "give_item")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_ADD_SECRET_REWARD, "add_secret_reward")
-ENUM_MAP_DEFINE(GF_SEQUENCE_EVENT_TYPE, GFS_PLAY_MUSIC,        "play_music")
-
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_TITLE, "title")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_NORMAL, "normal")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_CUTSCENE, "cutscene")
-ENUM_MAP_DEFINE(GF_LEVEL_TYPE, GFL_GYM, "gym")
-
-ENUM_MAP_DEFINE(GF_ACTION, GF_NOOP, "noop")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_GAME, "play_level")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_SAVED_GAME, "load_saved_game")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_CINE, "play_cutscene")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_DEMO, "play_demo")
-ENUM_MAP_DEFINE(GF_ACTION, GF_START_FMV, "play_fmv")
-ENUM_MAP_DEFINE(GF_ACTION, GF_EXIT_TO_TITLE, "exit_to_title")
-ENUM_MAP_DEFINE(GF_ACTION, GF_LEVEL_COMPLETE, "level_complete")
-ENUM_MAP_DEFINE(GF_ACTION, GF_EXIT_GAME, "exit_game")
-ENUM_MAP_DEFINE(GF_ACTION, GF_SELECT_GAME, "select_level")
